requirements.txt
# 核心Web框架
flask
# 跨域支持
flask-cors
# HTTP请求库
requests
# 读取.env配置文件
python-dotenv
# 可选：Cookie处理增强（非必需）
pycookiecheat